Data Ingestion/requirements.txt
chromadb
langdetect
requests
numpy
PyPDF2
pytesseract
Pillow
pdf2image
beautifulsoup4
lxml